% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{read Fatality Analysis Reporting System}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A valid file name to read the data from.}
}
\value{
This function returns the content of the read file in as a
        data frame.
}
\description{
This function reads a file from the Fatality Analysis Reporting
System and returns its content as a data frame.
}
\note{
This function will create an error if the file does not exist.
}
