% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{read multiple Fatality Analysis Reporting System files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Scalar year or vector of years to be read.}
}
\value{
This function returns a list (or a list of lists in the case
        that a vector of years is passed as a parameter) containing
        the data read from the corresponding files.
}
\description{
This function reads multiple files from the Fatality Analysis 
Reporting System. The files are organized in such a way that each
file contains the data of one year.
}
\examples{
fars_data <- fars_read_years(2013)
fars_data <- fars_read_years(c(2013, 2014))

}
