% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{create a summary of multiple Fatality Analysis Reporting System files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Scalar year or vector of years to be summarized.}
}
\value{
This function returns a data frame with a summary of the
        Fatality Analysis Reporting System data for the given years.
}
\description{
This function loads the relevant Fatality Analysis Reporting System
files and creates a summary of its data.
}
