% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make a valid file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year for which the file name should be created.}
}
\value{
This function returns a string containing the file name.
}
\description{
This function creates a valid Fatality Analysis Reporting System 
file name for a given year.
}
\examples{
fars_file_name <- make_filename(2013)

}
